testsite/requirements.txt
pip
Django<=5
django-decorator-include
django_extensions
django-debug-toolbar
-e ../